+++ changelog.md
@@ -3,6 +3,8 @@
 ### Misc
 - Added Headers to each bonus section on item sheet
 - Added Target Toggle to switch between any/all targets.
+- Better support for items in containers 
+  - Primarily when selecting Item/Weapon targets and when choosing what options are available for Weapon Type targets
 
 ### Bugfixes
 - Fixed remaining warning for rolls